+++ consensus/e2e_tests/utils_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/pokt-network/pocket/consensus"
 	typesCons "github.com/pokt-network/pocket/consensus/types"
-	mocksPer "github.com/pokt-network/pocket/persistence/types/mocks"
+	persistenceMocks "github.com/pokt-network/pocket/persistence/types/mocks"
 	"github.com/pokt-network/pocket/runtime"
 	"github.com/pokt-network/pocket/runtime/configs"
 	"github.com/pokt-network/pocket/runtime/defaults"
@@ -27,7 +27,6 @@ import (
 	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/shared/modules"
 	mockModules "github.com/pokt-network/pocket/shared/modules/mocks"
-	"github.com/pokt-network/pocket/shared/utils"
 	"github.com/pokt-network/pocket/state_machine"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
@@ -403,7 +402,6 @@ func basePersistenceMock(t *testing.T, _ modules.EventsChannel, bus modules.Bus,
 	ctrl := gomock.NewController(t)
 	persistenceMock := mockModules.NewMockPersistenceModule(ctrl)
 	persistenceReadContextMock := mockModules.NewMockPersistenceReadContext(ctrl)
-	blockStoreMock := mocksPer.NewMockKVStore(ctrl)
 
 	persistenceMock.EXPECT().GetModuleName().Return(modules.PersistenceModuleName).AnyTimes()
 	persistenceMock.EXPECT().Start().Return(nil).AnyTimes()
@@ -411,6 +409,7 @@ func basePersistenceMock(t *testing.T, _ modules.EventsChannel, bus modules.Bus,
 	persistenceMock.EXPECT().NewReadContext(gomock.Any()).Return(persistenceReadContextMock, nil).AnyTimes()
 	persistenceMock.EXPECT().ReleaseWriteContext().Return(nil).AnyTimes()
 
+	blockStoreMock := persistenceMocks.NewMockKVStore(ctrl)
 	blockStoreMock.EXPECT().Get(gomock.Any()).DoAndReturn(func(height []byte) ([]byte, error) {
 		heightInt := utils.HeightFromBytes(height)
 		if bus.GetConsensusModule().CurrentHeight() < heightInt {
@@ -418,8 +417,22 @@ func basePersistenceMock(t *testing.T, _ modules.EventsChannel, bus modules.Bus,
 		}
 		return codec.GetCodec().Marshal(testBlocks.getBlock(heightInt))
 	}).AnyTimes()
-
-	persistenceMock.EXPECT().GetBlockStore().Return(blockStoreMock).AnyTimes()
+	blockStoreMock.
+		// NB: The business logic in this mock and below is vital for testing state-sync end-to-end
+		EXPECT().
+		GetBlock(gomock.Any()).
+		DoAndReturn(func(height uint64) (*coreTypes.Block, error) {
+			if bus.GetConsensusModule().CurrentHeight() < height {
+				return nil, fmt.Errorf("requested height is higher than current height of the node's consensus module")
+			}
+			blockWithHeight := &coreTypes.Block{
+				BlockHeader: &coreTypes.BlockHeader{
+					Height: height,
+				},
+			}
+			return blockWithHeight, nil
+		}).
+		AnyTimes()
 
 	persistenceReadContextMock.EXPECT().GetMaximumBlockHeight().DoAndReturn(func() (uint64, error) {
 		// Check that we are retrieving a block at a height that was mocked by our test suite
@@ -430,17 +443,50 @@ func basePersistenceMock(t *testing.T, _ modules.EventsChannel, bus modules.Bus,
 		return 0, nil
 	}).AnyTimes()
 
-	persistenceReadContextMock.EXPECT().GetMinimumBlockHeight().DoAndReturn(func() (uint64, error) {
-		// mock minimum block height in persistence module to 1 if current height is equal or more than 1, else return 0 as the minimum height
-		if bus.GetConsensusModule().CurrentHeight() >= 1 {
-			return 1, nil
-		}
-		return 0, nil
-	}).AnyTimes()
+	persistenceMock.
+		EXPECT().
+		GetBlockStore().
+		Return(blockStoreMock).
+		AnyTimes()
 
-	persistenceReadContextMock.EXPECT().GetAllValidators(gomock.Any()).Return(bus.GetRuntimeMgr().GetGenesis().Validators, nil).AnyTimes()
-	persistenceReadContextMock.EXPECT().GetBlockHash(gomock.Any()).Return("", nil).AnyTimes()
-	persistenceReadContextMock.EXPECT().Release().AnyTimes()
+	persistenceReadContextMock.
+		EXPECT().
+		GetMaximumBlockHeight().
+		DoAndReturn(func() (uint64, error) {
+			height := bus.GetConsensusModule().CurrentHeight()
+			return height, nil
+		}).
+		AnyTimes()
+
+	persistenceReadContextMock.
+		EXPECT().
+		GetMinimumBlockHeight().
+		DoAndReturn(func() (uint64, error) {
+			// mock minimum block height in persistence module to 1 if current height is equal or more than 1, else return 0 as the minimum height
+			if bus.GetConsensusModule().CurrentHeight() >= 1 {
+				return 1, nil
+			}
+			return 0, nil
+		}).
+		AnyTimes()
+
+	persistenceReadContextMock.
+		EXPECT().
+		GetAllValidators(gomock.Any()).
+		Return(bus.GetRuntimeMgr().GetGenesis().Validators, nil).
+		AnyTimes()
+
+	persistenceReadContextMock.
+		EXPECT().
+		GetBlockHash(gomock.Any()).
+		Return("", nil).
+		AnyTimes()
+
+	persistenceReadContextMock.
+		EXPECT().
+		Release().
+		AnyTimes()
+
 	persistenceReadContextMock.EXPECT().GetValidatorExists(gomock.Any(), gomock.Any()).Return(true, nil).AnyTimes()
 
 	return persistenceMock
@@ -635,16 +681,10 @@ func WaitForNextBlock(
 	advanceTime(t, clck, 10*time.Millisecond)
 
 	blockStore := pocketNodes[1].GetBus().GetPersistenceModule().GetBlockStore()
-	heightBytes := utils.HeightToBytes(height)
-
-	blockBytes, err := blockStore.Get(heightBytes)
+	block, err := blockStore.GetBlock(height)
 	require.NoError(t, err)
 
-	var block coreTypes.Block
-	err = codec.GetCodec().Unmarshal(blockBytes, &block)
-	require.NoError(t, err)
-
-	return &block
+	return block
 }
 
 func waitForProposalMsgs(