+++ consensus/state_sync/server.go
@@ -4,10 +4,7 @@ import (
 	"fmt"
 
 	typesCons "github.com/pokt-network/pocket/consensus/types"
-	"github.com/pokt-network/pocket/shared/codec"
-	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
-	"github.com/pokt-network/pocket/shared/utils"
 )
 
 // This module is responsible for handling requests and business logic that advertises and shares
@@ -76,10 +73,11 @@ func (m *stateSync) HandleGetBlockRequest(blockReq *typesCons.GetBlockRequest) {
 	}
 
 	// get block from the persistence module
-	block, err := m.getBlockAtHeight(blockReq.Height)
+	blockStore := m.GetBus().GetPersistenceModule().GetBlockStore()
+	block, err := blockStore.GetBlock(blockReq.Height)
 	if err != nil {
-		m.logger.Err(err).Msg("Error getting block")
-		return
+		m.logger.Error().Err(err).Msgf("failed to get block at height %d", blockReq.Height)
+		return err
 	}
 
 	stateSyncMessage := typesCons.StateSyncMessage{
@@ -96,24 +94,4 @@ func (m *stateSync) HandleGetBlockRequest(blockReq *typesCons.GetBlockRequest) {
 		m.logger.Err(err).Msg("Error sending state sync message")
 		return
 	}
-}
-
-// Get a block from persistence module given block height
-func (m *stateSync) getBlockAtHeight(blockHeight uint64) (*coreTypes.Block, error) {
-	blockStore := m.GetBus().GetPersistenceModule().GetBlockStore()
-	heightBytes := utils.HeightToBytes(blockHeight)
-
-	blockBytes, err := blockStore.Get(heightBytes)
-	if err != nil {
-		m.logger.Error().Err(typesCons.ErrConsensusMempoolFull).Msg(typesCons.DisregardHotstuffMessage)
-		return nil, err
-	}
-
-	var block coreTypes.Block
-	err = codec.GetCodec().Unmarshal(blockBytes, &block)
-	if err != nil {
-		return &coreTypes.Block{}, err
-	}
-
-	return &block, nil
 }