+++ p2p/types/router.go
@@ -1,6 +1,6 @@
 package types
 
-//go:generate mockgen -source=$GOFILE -destination=./mocks/network_mock.go -package=mock_types github.com/pokt-network/pocket/p2p/types Network
+//go:generate mockgen -package=mock_types -destination=./mocks/network_mock.go github.com/pokt-network/pocket/p2p/types Router,RouterConfig
 
 import (
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
@@ -25,4 +25,10 @@ type Router interface {
 	// Handles the raw data received from the network and returns the data to be processed
 	// by the application layer.
 	HandleNetworkData(data []byte) ([]byte, error)
+}
+
+// RouterConfig is used to configure `Router` implementations and to test a
+// given configuration's validity.
+type RouterConfig interface {
+	IsValid() error
 }