+++ p2p/utils/logging.go
@@ -4,10 +4,9 @@ import (
 	"net"
 
 	"github.com/libp2p/go-libp2p/core/network"
-	"github.com/rs/zerolog"
-
 	"github.com/pokt-network/pocket/p2p/types"
 	"github.com/pokt-network/pocket/shared/modules"
+	"github.com/rs/zerolog"
 )
 
 type scopeCallback func(scope network.ResourceScope) error