+++ persistence/context.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/pokt-network/pocket/persistence/blockstore"
 	"github.com/pokt-network/pocket/persistence/indexer"
-	"github.com/pokt-network/pocket/persistence/kvstore"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/modules"
 )
@@ -29,7 +29,7 @@ type PostgresContext struct {
 	stateHash string
 	// TECHDEBT(#361): These three values are pointers to objects maintained by the PersistenceModule.
 	//                 Need to simply access them via the bus.
-	blockStore kvstore.KVStore
+	blockStore blockstore.BlockStore
 	txIndexer  indexer.TxIndexer
 	stateTrees *stateTrees
 
@@ -68,8 +68,8 @@ func (p *PostgresContext) Commit(proposerAddr, quorumCert []byte) error {
 		return err
 	}
 
-	// Store block in the KV store
-	if err := p.storeBlock(block); err != nil {
+	// Save the block in the BlockStore at the current height
+	if err := p.blockStore.StoreBlock(uint64(p.Height), block); err != nil {
 		return err
 	}