+++ persistence/debug.go
@@ -5,10 +5,7 @@ import (
 	"runtime/debug"
 
 	"github.com/pokt-network/pocket/persistence/types"
-	"github.com/pokt-network/pocket/shared/codec"
-	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/messaging"
-	"github.com/pokt-network/pocket/shared/utils"
 	"github.com/pokt-network/smt"
 )
 
@@ -45,15 +42,9 @@ func (m *persistenceModule) HandleDebugMessage(debugMessage *messaging.DebugMess
 func (m *persistenceModule) showLatestBlockInStore(_ *messaging.DebugMessage) {
 	// TODO: Add an iterator to the `kvstore` and use that instead
 	height := m.GetBus().GetConsensusModule().CurrentHeight() - 1
-	blockBytes, err := m.GetBlockStore().Get(utils.HeightToBytes(height))
+	block, err := m.GetBlockStore().GetBlock(height)
 	if err != nil {
 		m.logger.Error().Err(err).Uint64("height", height).Msg("Error getting block from block store")
-		return
-	}
-
-	block := &coreTypes.Block{}
-	if err := codec.GetCodec().Unmarshal(blockBytes, block); err != nil {
-		m.logger.Error().Err(err).Uint64("height", height).Msg("Error decoding block from block store")
 		return
 	}