+++ rpc/module.go
@@ -1,11 +1,8 @@
 package rpc
 
 import (
-
-	// importing because used by code-generated files that are git ignored and to allow go mod tidy and go mod vendor to function properly
 	_ "github.com/getkin/kin-openapi/openapi3"
 	_ "github.com/labstack/echo/v4"
-
 	"github.com/pokt-network/pocket/logger"
 	"github.com/pokt-network/pocket/runtime/configs"
 	"github.com/pokt-network/pocket/shared/modules"