+++ runtime/manager_test.go
@@ -4208,11 +4208,11 @@ func TestNewManagerFromReaders(t *testing.T) {
 						HealthCheckPeriod: "30s",
 					},
 					P2P: &configs.P2PConfig{
-						PrivateKey:      "0ca1a40ddecdab4f5b04fa0bfed1d235beaa2b8082e7554425607516f0862075dfe357de55649e6d2ce889acf15eb77e94ab3c5756fe46d3c7538d37f27f115e",
-						Hostname:        "node1.consensus",
-						Port:            defaults.DefaultP2PPort,
-						ConnectionType:  configTypes.ConnectionType_TCPConnection,
-						MaxMempoolCount: 1e5,
+						PrivateKey:     "0ca1a40ddecdab4f5b04fa0bfed1d235beaa2b8082e7554425607516f0862075dfe357de55649e6d2ce889acf15eb77e94ab3c5756fe46d3c7538d37f27f115e",
+						Hostname:       "node1.consensus",
+						Port:           defaults.DefaultP2PPort,
+						ConnectionType: configTypes.ConnectionType_TCPConnection,
+						MaxNonces:      1e5,
 					},
 					Telemetry: &configs.TelemetryConfig{
 						Enabled:  true,
@@ -4240,7 +4240,7 @@ func TestNewManagerFromReaders(t *testing.T) {
 			},
 		},
 		{
-			name: "unset MaxMempoolCount should fallback to default value",
+			name: "unset MaxNonces should fallback to default value",
 			args: args{
 				configReader: strings.NewReader(string(`{
 					"p2p": {
@@ -4256,11 +4256,11 @@ func TestNewManagerFromReaders(t *testing.T) {
 			want: &Manager{
 				config: &configs.Config{
 					P2P: &configs.P2PConfig{
-						PrivateKey:      "4ff3292ff14213149446f8208942b35439cb4b2c5e819f41fb612e880b5614bdd6cea8706f6ee6672c1e013e667ec8c46231e0e7abcf97ba35d89fceb8edae45",
-						Hostname:        "node1.consensus",
-						Port:            42069,
-						ConnectionType:  configTypes.ConnectionType_TCPConnection,
-						MaxMempoolCount: defaults.DefaultP2PMaxMempoolCount,
+						PrivateKey:     "4ff3292ff14213149446f8208942b35439cb4b2c5e819f41fb612e880b5614bdd6cea8706f6ee6672c1e013e667ec8c46231e0e7abcf97ba35d89fceb8edae45",
+						Hostname:       "node1.consensus",
+						Port:           42069,
+						ConnectionType: configTypes.ConnectionType_TCPConnection,
+						MaxNonces:      defaults.DefaultP2PMaxNonces,
 					},
 					Keybase: defaultCfg.Keybase,
 				},