+++ consensus/doc/CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.51] - 2023-05-05
+## [0.0.0.53] - 2023-05-25
 
 - Added state sync channels `blocksReceived` and `metadataReceived`
 - Implemented `blockApplicationLoop()`
 - Implemented state sync module functions `Start()` and `Stop()`
 - Implemented `WaitForNetworkFSMEvents()` function in test utils
+
+## [0.0.0.52] - 2023-05-24
+
+- Updates consensus tests to use GetBlock mock instead of Get mock
+
+## [0.0.0.51] - 2023-05-16
+
+- Updates consensus test to use BlockStore mock instead of KVStore mock
 
 ## [0.0.0.50] - 2023-05-03